Cleaned up various files, fixed various gcc warnings. Removed nError.
Fixed nginx pointer size conversion warning.
Fixed various implicitly declared functions.
Cleaned up FileSystem class, removing callback functions.

+++ ext/nginx/src/ngx_http_webapp_module.c
@@ -229,7 +229,7 @@ static void conn_write(ngx_event_t *ev) {
 						for(cl = r->request_body->bufs; cl; cl = cl->next) {
 							int current_buf_len;
 							buf = cl->buf;
-							current_buf_len = (int)buf->last - (int)buf->pos;
+							current_buf_len = (int)(buf->last - buf->pos);
 							req = ngx_copy(req, cl->buf->pos, current_buf_len);
 							bytes_out += current_buf_len;
 						}
